requirements.txt
Flask
twilio
openai
python-dotenv
